+++ libs/extension/src/background/ws/mod.rs
@@ -1,3 +1,5 @@
+mod channel;
 pub mod on_message;
 pub mod websocket;
+
 pub use on_message::*;
